+++ src/index.mjs
@@ -1,0 +1,4 @@
+// ESM wrapper
+import a from './getFlags.js';
+
+export default a;