+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.1.0-alpha.0 (2025-07-03)
+
+- feat: optimize npm package for smaller size: 12kB -> 5kB
+- feat: add d.ts file
+- refactor: refactor code, add more comments
+- chore: add rollup to build clean npm package
+- docs: improve readme
+
+## 0.0.2 (2025-07-02)
+
+- fix: resolving files in package.json for CommonJS
+
 ## 0.0.1 (2025-07-01)
 
 - feat: create small and fast CLI flag parser with supports of all base formats